modules structure ported to the latest master state
