+++ modules/dry_adiabat_setup.py
@@ -1,0 +1,34 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Jan 23 12:06:35 2023
+
+@authors: 
+Tim Lichtenberg (TL)    
+Ryan Boukrouche (RB)
+"""
+
+import numpy as np
+import utils.GeneralAdiabat as ga # Moist adiabat with multiple condensibles
+    
+# Dry adiabat profile
+def dry_adiabat_atm(atm):
+
+    R_universal = 8.31446261815324              # Universal gas constant, J.K-1.mol-1
+
+    # Calculate cp from molar concentrations
+    cp_mix = 0.
+    for vol in atm.vol_list.keys():
+        cp_mix += atm.vol_list[vol] * ga.cpv(vol, atm.ts)
+
+    # Calculate dry adiabat slope
+    atm.Rcp = R_universal / cp_mix
+
+    # Calculate dry adiabat temperature profile for staggered nodes (from ground up)
+    for idx, prsl in enumerate(atm.pl):
+        atm.tmpl[idx] = atm.ts * ( prsl / atm.ps ) ** ( atm.Rcp )
+
+    # Interpolate temperature from staggered nodes
+    atm.tmp = np.interp(atm.p, atm.pl, atm.tmpl)
+
+    return atm
