+++ modules/dry_adjustment.py
@@ -1,0 +1,51 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Jan 23 11:36:49 2023
+
+@authors:    
+Ryan Boukrouche (RB)
+"""
+
+# Dry convective adjustment
+def DryAdj(atm):
+
+    T   = atm.tmp
+    p   = atm.p
+    
+    # Rcp is global
+    # Downward pass
+    for i in range(len(T)-1):
+        T1,p1 = T[i],p[i]
+        T2,p2 = T[i+1],p[i+1]
+        
+        # Adiabat slope
+        pfact = (p1/p2)**atm.Rcp
+        
+        # If slope is shallower than adiabat (unstable), adjust to adiabat
+        if T1 < T2*pfact:
+            Tbar = .5*(T1+T2) # Equal layer masses
+                              # Not quite compatible with how
+                              # heating is computed from flux
+            T2 = 2.*Tbar/(1.+pfact)
+            T1 = T2*pfact
+            atm.tmp[i]   = T1
+            atm.tmp[i+1] = T2
+    
+    # Upward pass
+    for i in range(len(T)-2,-1,-1):
+
+        T1,p1 = T[i],p[i]
+        T2,p2 = T[i+1],p[i+1]
+        pfact = (p1/p2)**atm.Rcp
+
+        if T1 < T2*pfact:
+            Tbar = .5*(T1+T2) # Equal layer masses
+                              # Not quite compatible with how
+                              # heating is computed from flux
+            T2 = 2.*Tbar/(1.+pfact)
+            T1 = T2*pfact
+            atm.tmp[i]   = T1
+            atm.tmp[i+1] = T2 
+
+    return atm
