+++ modules/spectral_planck_surface.py
@@ -1,0 +1,24 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Jan 23 12:16:20 2023
+
+@authors: 
+Tim Lichtenberg (TL)    
+Ryan Boukrouche (RB)
+"""
+
+import numpy as np
+
+def surf_Planck_nu(atm):
+    h   = 6.63e-34
+    c   = 3.0e8
+    kb  = 1.38e-23
+    B   = np.zeros(len(atm.band_centres))
+    c1  = 1.191042e-5
+    c2  = 1.4387752
+    for i in range(len(atm.band_centres)):
+        nu      = atm.band_centres[i]
+        B[i]    = (c1*nu**3 / (np.exp(c2*nu/atm.ts)-1))
+    B   = (1.-atm.albedo_s) * np.pi * B * atm.band_widths/1000.0
+    return B
